+++ src/ui/public/courier/fetch/is_request.js
@@ -1,4 +1,4 @@
-import AbstractRequestProvider from './request';
+import AbstractRequestProvider from '../fetch_types/abstract_request';
 
 export default function IsRequestProvider(Private) {
   const AbstractRequest = Private(AbstractRequestProvider);
