+++ src/ui/public/courier/fetch_types/__tests__/es_search_strategy.js
