+++ src/ui/public/courier/fetch_types/__tests__/segmented.js
