+++ src/ui/public/courier/fetch_types/__tests__/segmented_index_selection.js
