+++ src/ui/public/courier/looper/doc.js
@@ -1,18 +1,18 @@
 import FetchProvider from '../fetch';
 import LooperProvider from './_looper';
-import DocStrategyProvider from '../fetch/strategy/doc';
+import EsDocStrategyProvider from '../fetch_types/es_doc_strategy';
 
 export default function DocLooperService(Private) {
   var fetch = Private(FetchProvider);
   var Looper = Private(LooperProvider);
-  var DocStrategy = Private(DocStrategyProvider);
+  var esDocStrategy = Private(EsDocStrategyProvider);
 
   /**
    * The Looper which will manage the doc fetch interval
    * @type {Looper}
    */
   var docLooper = new Looper(1500, function () {
-    fetch.fetchQueued(DocStrategy);
+    fetch.fetchQueued(esDocStrategy);
   });
 
   return docLooper;
